Origin use registered prefix when polling broker
When the Director ConnectToService it passes the registered prefix of
the Origin to connect to it. When the Origin polls the broker it
makes two polls, one for the web endpoints and the other for xrootd. The
problem lies in that the xrootd poll was using its own xrootd url, which
is unregistered, in the polling token. Instead, we should poll using the
registered prefix. This way, both the director and origin (web and
xrootd) can all talk on the same channel.