Remove 'Content-Disposition: attachment' header from Origins
This header is no longer needed by the web client, and it prevents
people from browsing a directory tree at their Origins (erm... a...
collections tree? Whatever, I'm talking about a POSIX Origin where
it really _is_ a directory). By removing this, the directory tree
is rendered as a web page instead of downloaded as HTML.