Images: Pin xrdhttp-pelican to a specific version
This is necessary because xrdhttp-pelican is locked to a specific minor
version of xrootd and we're pinning the xrootd version, so we can't just
take whatever is the newest in the repos.

Take it from Koji if we're not building it ourselves.

(cherry picked from commit ac241c39cc6804a9c3152c841723d9c21f8981e3)