Reconfigure global fed info object during fed tests
This was a subtle bug I uncovered while trying to write a unit test
that distinguished between the fed metadata hosted by the discovery
server and the test's Director.

It turned out that the duplicate metadata hosted by the test Director
wasn't matching the discovery server's because it wasn't enough to
set the Discovery URL as a param for fed tests because by that point,
global fed discovery had already run by the Director (who thought it
was the fed root when it started -- a bootstrapping issue).

To achieve the desired behavior, I also needed to overwrite the global
fed info to reflect what the Director would have discovered had it
used this as its source of truth.

I don't think this is actually a bug in the code, only in the awkward
approach we have to setting up federations for tests.