Add param that disables Director's hosting of federation metadata
There is an entire class of bugs that revolves around having the
Director host a copy of federation metadata when it isn't also the
root discovery URL for the federation.

Adding this knob doesn't _solve_ those bugs, but it should help us
detect them more easily by allowing us to turn of Director metadata
when we don't want people using the Director for that purpose.