Add metric to detect attempts to get fed info from Director
This was requested by the Ops team to smooth the transition toward
turning off metadata hosting at the Director in OSDF.

This commit adds a metric we can use to watch who is still hitting
the Director for fed metadata. I tried to give the metric enough
contextual info to determine "who" is asking for the data by recording
a masked address and a service type (bootstrapped from the request's
user agent).

Importantly, this changed my original feature design slightly -- rather
than completely turning off the metadata hosting endpoint by de-registering
it with gin, this keeps it registered so we can collect metrics. If
the metadata hosting is disabled, we return a `410 Gone` as soon as
we record who tried to access the info.